requirements.txt
python-telegram-bot>=12.7
requests==2.24.0
six==1.15.0
psycopg2>=2.7.3.1
jsonpickle >=0.8

telethon>=1.14
pillow>=6.1.0
cryptg>=0.2
aiohttp>=3.6.0
hachoir>=3.0a5

kafka-python===2.0.1
aiokafka>=0.6.0
pika>=1.1.0
aio-pika>=6.6.1
redis>=3.5.3

asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dnspython==1.16.0
idna==2.9
markdown-strings==3.3.0
pymongo==3.10.1
pyTelegramBotAPI==3.7.3
pytz==2020.1
schedule==1.1.0
sqlparse==0.3.1
urllib3==1.25.9
beautifulsoup4==4.9.3
pandas==1.2.4
discord==1.7.3
scipy==1.5.4